+++ spec/light/services/with_done_bypasses_always_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe WithDoneBypassesAlways do
   it "bypasses always steps when done! is called" do
     service = described_class.run
 
-    expect(service).to be_success
+    expect(service).to be_successful
     expect(service.trace).to eq([:work, :after_done])
   end
 end